+++ Braintree/CreditCardOptionsRequest.cs
@@ -10,6 +10,7 @@ namespace Braintree
     {
         public Boolean? VerifyCard { get; set; }
         public Boolean? MakeDefault { get; set; }
+        public Boolean? FailOnDuplicatePaymentMethod { get; set; }
         public String VerificationMerchantAccountId { get; set; }
         public String UpdateExistingToken { get; set; }
 
@@ -30,6 +31,7 @@ namespace Braintree
             builder.AddElement("make-default", MakeDefault);
             builder.AddElement("verification-merchant-account-id", VerificationMerchantAccountId);
             builder.AddElement("verify-card", VerifyCard);
+            builder.AddElement("fail-on-duplicate-payment-method", FailOnDuplicatePaymentMethod);
             builder.AddElement("update-existing-token", UpdateExistingToken);
 
             return builder;