+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.14.0
+
+* Adds search for duplicate credit cards given a payment method token
+* Adds flag to fail saving credit card to vault if card is duplicate
+* Changes `internal` functions to `protected internal` functions for easier unit testing
+
 ## 2.13.4
 
 * Exposes plan_id on transactions